refactor: RequestWithSession, ContainerSession 타입을 별도 파일로 분리
