+++ backend/src/common/types/request.ts
@@ -1,0 +1,6 @@
+import { ContainerSession } from "./session";
+import { Request } from "express";
+
+export interface RequestWithSession extends Request {
+    session: ContainerSession;
+}
