+++ backend/src/common/types/session.ts
@@ -1,0 +1,5 @@
+export interface ContainerSession {
+    containerId: string;
+    startTime: Date;
+    renew: boolean;
+}
